Async is a reserved keyword starting in Python 3.7
Which means using it as a variable or attribute name is a syntax error.
Since KonText claims to be compatible with Python 3.6+, we might as well
future-proof it right now :)

I changed `async` to `asnc` because this renaming was already used in
some places in the codebase; another option would be `async_`.