Rewrite internal http client lifecycle.
Now it is a single session per each request. This means that in case
a single KonText request requires to perform multiple http requests
to different services (this is typical e.g. for some plug-in requests),
only a single session is used for all the requests.

The ClientSession is attached to request.ctx and is initialized
via Sanic request middleware and closed via Sanic response middleware.