Rewrite the way how switching to a different "main" corpus is performed:
Now it is based on rewriting the initial query and appending the
"switch main corp" operation. Everything is controlled by the query
replay model.

This fixed different problems of the previous solution (e.g. unability
to open operations in the "breadcrumb" navigation). It also produces
more compact (and correct) CQL op. sequence.